Configure Jest for unit/integration test separation
- Reorganize tests into tests/unit/ and tests/integration/ directories
- Move hyco-https tests to hyco-https/tests/integration/
- Move test-utils tests to test-utils/tests/unit/
- Update jest.config.js with explicit testMatch patterns
- Add test:unit and test:integration npm scripts
- Tests can be filtered via --testPathPattern=unit or --testPathPattern=integration
